test: 리뷰 반영 stringCalculator 단위 테스트
